backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1
anthropic==0.18.1
pymongo==4.6.2
python-dotenv==1.0.1
nltk==3.8.1
requests==2.31.0
